tests/wpt/web-platform-tests/uievents/order-of-events/mouse-events/click-cancel-manual.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Default action is canceled if the click event is canceled.</title>
    <link rel="help" href="https://w3c.github.io/uievents/#events-clickevent-event-order">
    <meta name="flags" content="interact">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <ol>
      <li>Check the following item.</li>
      <li>Hover over the button.</li>
    </ol>
    <input type="checkbox" id="checkbox">Check me!<br>
    <button type="button" id="done">Button</button>
</body>
<script>
setup({explicit_timeout: true});

window.onload = function () {
  var box = document.getElementById("checkbox");
  box.addEventListener("click", function(event) {
    event.preventDefault();
  })

  async_test(function(t) {
    document.getElementById("done").addEventListener("mouseover", function () {
      t.step(function () {
        assert_false(box.checked);
        t.done();
      });
    }, false);
  }, "Default event is canceled if the click event is canceled");
};
    </script>
</html>
